whatsgranny/app/requirements.txt
aiohttp-retry==2.8.3 
aiohttp==3.8.4 
aiosignal==1.3.1 
anyio==3.6.2 
arrow==1.2.3 
async-timeout==4.0.2 
attrs==23.1.0 
bleach==6.0.0 
cachetools==5.3.0 
certifi==2022.12.7
cffi==1.15.1 
charset-normalizer==3.1.0
click-log==0.4.0 
click==8.1.3 
cmake==3.26.3
colorama==0.4.6 
cryptography==40.0.2
debugpy==1.6.7 
deprecation==2.1.0 
docutils==0.19 
dotty-dict==1.3.1 
ffmpeg-python==0.2.0 
filelock==3.12.0 
flask==2.2.3 
frozenlist==1.3.3 
future==0.18.3 
gitdb==4.0.10 
gitpython==3.1.31 
google-api-core==2.11.0 
google-auth==2.17.3 
google-cloud-core==2.3.2 
google-cloud-storage==2.8.0 
google-cloud==0.34.0 
google-crc32c==1.5.0 
google-resumable-media==2.4.1 
googleapis-common-protos==1.59.0 
gotrue==1.0.1 
gunicorn==20.1.0 
h11==0.14.0 
httpcore==0.16.3 
httpx==0.23.3 
huggingface-hub==0.13.4 
idna==3.4 ; python_version >= "3.11"
importlib-metadata==6.5.1 
invoke==1.7.3 
itsdangerous==2.1.2 
jaraco-classes==3.2.3 
jeepney==0.8.0 
jinja2==3.1.2 
keyring==23.13.1 
lit==16.0.1 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.11" and python_version < "4.0" 
markupsafe==2.1.2 
more-itertools==9.1.0 
mpmath==1.3.0 
multidict==6.0.4 
mysql-connector-python==8.0.33 
networkx==3.1 
numpy==1.24.2 
nvidia-cublas-cu11==11.10.3.66 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.11" and python_version < "4.0" 
nvidia-cuda-cupti-cu11==11.7.101 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.11" and python_version < "4.0" 
nvidia-cuda-nvrtc-cu11==11.7.99 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.11" and python_version < "4.0" 
nvidia-cuda-runtime-cu11==11.7.99 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.11" and python_version < "4.0" 
nvidia-cudnn-cu11==8.5.0.96 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.11" and python_version < "4.0" 
nvidia-cufft-cu11==10.9.0.58 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.11" and python_version < "4.0" 
nvidia-curand-cu11==10.2.10.91 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.11" and python_version < "4.0" 
nvidia-cusolver-cu11==11.4.0.1 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.11" and python_version < "4.0" 
nvidia-cusparse-cu11==11.7.4.91 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.11" and python_version < "4.0" 
nvidia-nccl-cu11==2.14.3 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.11" and python_version < "4.0" 
nvidia-nvtx-cu11==11.7.91 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.11" and python_version < "4.0" 
openai==0.27.4 
packaging==23.1 
pillow==9.5.0 
pkginfo==1.9.6 
postgrest==0.10.6 
protobuf==3.20.3 
pyasn1-modules==0.3.0 
pyasn1==0.5.0 
pycparser==2.21 
pydantic==1.10.7 
pygments==2.15.1 
pyjwt==2.6.0 
python-dateutil==2.8.2 
python-dotenv==1.0.0 
python-gitlab==3.14.0 
python-semantic-release==7.33.2 
pytz==2023.3 
pywin32-ctypes==0.2.0
pyyaml==6.0 
rapidfuzz==3.0.0 
readme-renderer==37.3 
realtime==1.0.0 
redis-sentinel-url==1.0.1 
redis==4.5.4 
regex==2023.3.23 
reportlab==3.6.12
requests-toolbelt==0.10.1 
requests==2.28.2 
rfc3986==1.5.0 
rfc3986[idna2008]==1.5.0 
rq-dashboard @ git+https://github.com/jace/rq-dashboard.git@d506c6b43545e502673b32ed70a6d89d180ac324 
rq==1.13.0 
rsa==4.9 
secretstorage==3.3.3
semver==2.13.0 
setuptools==67.7.1 
six==1.16.0 
smmap==5.0.0 
sniffio==1.3.0 
storage3==0.5.2 
strenum==0.4.10 
supabase==1.0.3 
supafunc==0.2.2 
sympy==1.11.1 
tokenizers==0.13.3 
tomlkit==0.11.7 
torch==2.0.0 
tqdm==4.65.0 
transformers==4.28.1 
twilio==8.1.0 
twine==3.8.0 
typing-extensions==4.5.0 
urllib3==1.26.15 
webencodings==0.5.1 
websockets==10.4 
werkzeug==2.2.3 
wheel==0.40.0 
whisper-openai==1.0.0 
yarl==1.9.1 
zipp==3.15.0 
